Disabling per-frame gl.getError()
In theory clearing the GL error at the beginning of the frame
is good for debugging, and shouldn't actually harm anything (possibly stall
as it's the first GL call of the frame, but theoretically not much more than
whatever the next GL call would be), however in practice this is adding up
to a couple ms (when running at /max_fps 1000) in Chrome.  Does not seem to
have any effect either way under GPU-bound conditions though.